refactor: externalize Footer component
